- Make the playback and volume controllable by keyboard.
All players have some sort of keyboard control, Libreflix should
have this too! This patch makes the playback controllable by
the left and right arrow keys and the volume controllable by
the up and down keys.

The volumeChange function was rewritten to utilize the
changevolumeDelta so we don't have code duplication!