- Prevent EXDEV errors and some typos corrections.
fs.rename fails on cross device files.
So if the the tmpfs is mounted on /tmp (which is to be the default)
and the /tmp itself is on other device the call for fs.rename will
trigger an [Error: EXDEV: cross-device link not permitted...]

This make sure that we always have the correct behaviour no matter
the actual underlying devices that we're on.

The call of fs.remove is kind of optional, because by default
the contents of tmpfs is cleared upon reinitialization. But to
ensure that the disk space is preserved we're calling it here.

Please beware of this and remove the fs.remove call if any performance
degradation is perceived at all.

Other thing to keep in mind is that those functions aren't default
fs module calls, but instead fs-extra calls.

Reference:
 https://github.com/jprichardson/node-fs-extra